MacOS Setup Script Fix (#32)
#### Description
If a user has MacOS and has not installed `coreutils` through `brew` they will not have `realpath`. This PR introduces a workaround for this - but at the cost of it no longer follows symlinks. This is probably okay - and the user has to provide confirmation to the path given to them.
#### Tests
Tested on MacOS and Git Bash
#### Documentation
N/A
#### Primary reviewers
If your change will take a long time to review, you can name at most two primary reviewers who are ultimately responsible for reviewing this request. @ mention them.
@jared-improbable @jessicafalk